// src/app/models/campaigns.module.ts
export class CampaignsModule {
    CampaignName: string;
    Email: string;
    CampaignDescription: string
    SubjectLine: string
    FromName: string;
    ToMail: string;
    camp_ID: number;
    Temp_ID: string;
    TargetListIds:any
    Status: string;
    Cc: string;
    Bcc: string;
    FromEmailAddress: string;
    ReplyToEmailAddress: string;
    UnsubscribeInfo: string;
    userId: any;
    ImagePath: string;
    SelectCondition: string;
    APISelectCondition: string;
    UserType: any;
    HTMLContent: string;
    TextContent: string;
    IsActive: string;
    IsTrigger: string;
    UsedTrigger: string;
    SendType: string;
    ScheduledDatetime: Date;
    CreatedDate: Date;
    ModifiedDate: Date;
    CreatedBy: any;
    ModifiedBy: any;
    campAPI_ID: any;
    totalRecipient: any;
    status:any
    number_of_contacts:any

    cordsCount: string;
    ContactTargeted: string;
    SuccessfullDelivered: string;
    Bounced: string;
    TotalOpens: string;
    TotalClicks: string;
    CampStatus: string;
    Radio: string;
    Template_name:any
}